+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - xcode 13.0 (ios 15.0)
 - android 12 (api 31)
 - dotnet sdk 5.0.400
+- fix java unhandled exceptions on android by porting DecompressInterceptor c# code to java
 - build : NuGetizer 0.7.5
 - test : Microsoft.Extensions.DependencyInjection 5.0.2
 - test : Serilog.Sinks.Console 4.0.0